### 2022-01-24 #4096 @kno10
> @kno10 – due to the small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week.

Thank you @arfon!
The majority of the code lives in the Rust sister repository (we maybe should have put the paper draft there - the python part really is just a wrapper, but will likely have many more users) - maybe you can run whedon statistics on that repository instead:
https://github.com/kno10/rust-kmedoids
and of course much of the scholarly effort previously went into the Java version - the Rust version was a pretty direct port of this (except for the new parallelization by @lale1009).
