### 2022-01-28 d392667
Fix "random" instead of "np.random". Closes#2
