### 2022-01-28 3ee5696
really fix random seeding when no seed is given
