### 2022-01-29 #4096 @arfon
Got it. Thanks @kno10. FYI JOSS editors, this is the `cloc` output for the [https://github.com/kno10/rust-kmedoids](https://github.com/kno10/rust-kmedoids) repo.

```
      19 text files.
      19 unique files.
       3 files ignored.

Language                     files          blank        comment           code
Rust                            14             72            472           1543
Markdown                         2             40              0             94
TOML                             1              7              1             38
SUM:                            17            119            473           1675
```
