### 2022-02-14 #4096 @TahiriNadia
> @timClicks, @TahiriNadia: Would you be interested in reviewing this submission to [The Journal of Open Source Software](https://joss.theoj.org/)? Reviews are open and based on a checklist. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you have any questions or concerns please let me know.

Yes, I will!
