### 2022-02-19 #4183 @mikldk
@timClicks, @TahiriNadia: Thanks for agreeing to review.

Please note that both the repos should be reviewed (<https://github.com/openjournals/joss-reviews/issues/4096#issuecomment-1020333611>):

* The engine: <https://github.com/kno10/rust-kmedoids>
* Python interface: <https://github.com/kno10/python-kmedoids>

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

@kno10: As you see both repos will be reviewed. This also means that when the repos are ready, you will have to prepare a combined archive (e.g. [zenodo.org](https://zenodo.org/)) of both repositories.
