### 2022-02-20 kno10/python-kmedoids#3 @timClicks
Requires installation instructions
Hello, I'm one of the reviewers for your JOSS submission[[1](https://github.com/openjournals/joss-reviews/issues/4183)] . The review guidelines require installation instructions in your documentation.

>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

As you're using `cargo` to build the project rather than tools which are more familiar to the Python community, perhaps you should explain how to run `cargo build --release` and then make `./target/release/libkmedoids.so` available as a Python module.
