### 2022-02-20 kno10/python-kmedoids#4 @timClicks
Integration tests
Hello, I'm one of the reviewers for your JOSS submission[https://github.com/openjournals/joss-reviews/issues/4183] .

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

As your submission makes use of two repositories, it makes sense that there some testing that they work together. This can either be description of how to functionality works as and/or automated integration tests.

There are also no tests within your Python module.
