### 2022-02-21 kno10/python-kmedoids#3 @lale1009
I included an installation instruction in the README.md. Hope that is fair enough. Thanks!
