### 2022-02-22 kno10/python-kmedoids#3 @kno10
@lale1009 we still need to figure out a cross-platform way to automate an installation into the current environment, e.g., via python setuptools.
