### 2022-02-22 a4a31d7
installation with maturin
