### 2022-02-22 kno10/python-kmedoids#3 @kno10
I updated the instructions to simply use `maturin`. So far, I had only been using this for building the pypi wheels, but it can also be used to build and install the package locally easily.
