### 2022-03-03 09338ac
Update manual installation instructions for integration tests
