+++ tests/test_integration.py
@@ -6,7 +6,7 @@ import unittest
 
 class Test_kmedoids(unittest.TestCase):
     def test_pam(self):
-        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]])
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)
         pam_build = kmedoids.pam_build(dist, 2);
         pam_build_rust = kmedoids.kmedoids._pam_build_i32(dist, 2);
         assert pam_build.loss == 11;
@@ -18,7 +18,7 @@ class Test_kmedoids(unittest.TestCase):
         assert np.array_equal(pam_swap.medoids, pam_swap_rust[2]);
 
     def test_fasterpam(self):
-        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]])
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)
         fp = kmedoids.fasterpam(dist, 2, init="build");
         par_fp = kmedoids.fasterpam(dist, 2, init="build", n_cpu = 2);
         fp_rust = kmedoids.kmedoids._fasterpam_i32(dist, fp.medoids, 100);
@@ -30,7 +30,7 @@ class Test_kmedoids(unittest.TestCase):
 
 
     def test_fastpam1(self):
-        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]])
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)
         fp1 = kmedoids.fastpam1(dist, 2, init="build");
         assert fp1.loss == 9;
         fp1_rust = kmedoids.kmedoids._fastpam1_i32(dist, fp1.medoids, 100);
@@ -38,14 +38,14 @@ class Test_kmedoids(unittest.TestCase):
         assert fp1.loss == fp1_rust[0];
 
     def test_alternating(self):
-        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]])
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)
         alt = kmedoids.alternating(dist, 2, init="build");
         alt_rust = kmedoids.kmedoids._alternating_i32(dist, alt.medoids, 100);
         assert np.array_equal(alt.medoids, alt_rust[2]);
         assert alt.loss == alt_rust[0];
 
     def test_silhouette(self):
[…]
