### 2022-03-03 kno10/python-kmedoids#4 @lale1009
Added integration tests for fasterpam, pam, fastpam1 and silhouette: 1748553b0e33169f9c6a3889e34b1c08e6d49552

Documentation updated accordingly: 09338ac85831d853d92ffe9608ce199aa5a94c95
