### 2022-03-07 d239257
Add interface analog to sklearn.
