+++ tests/test_integration.py
@@ -53,5 +53,33 @@ class Test_kmedoids(unittest.TestCase):
         assert sil == par_sil;
         assert sil == sil_rust[0];
 
+    def test_sklearn_interface(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32);
+        kmed = kmedoids.KMedoids(2, method='fasterpam', init='build');
+        res_sk = kmed.fit(dist);
+        res = kmedoids.fasterpam(dist, 2, init="build");
+        assert res_sk.inertia_ == res.loss;
+        assert np.array_equal(res_sk.labels_, res.labels);
+        assert np.array_equal(res_sk.medoid_indices_, res.medoids);
+        kmed = kmedoids.KMedoids(2, method='pam', init='build')
+        res_sk = kmed.fit(dist);
+        res = kmedoids.pam(dist, 2, init="build");
+        assert res_sk.inertia_ == res.loss;
+        assert np.array_equal(res_sk.labels_, res.labels);
+        assert np.array_equal(res_sk.medoid_indices_, res.medoids);
+        kmed = kmedoids.KMedoids(2, method='fastpam1', init='build');
+        res_sk = kmed.fit(dist);
+        res = kmedoids.fastpam1(dist, 2, init="build");
+        assert res_sk.inertia_ == res.loss;
+        assert np.array_equal(res_sk.labels_, res.labels);
+        assert np.array_equal(res_sk.medoid_indices_, res.medoids);
+        kmed = kmedoids.KMedoids(2, method='alternate', init='build');
+        res_sk = kmed.fit(dist);
+        res = kmedoids.alternating(dist, 2, init="build");
+        assert res_sk.inertia_ == res.loss;
+        assert np.array_equal(res_sk.labels_, res.labels);
+        assert np.array_equal(res_sk.medoid_indices_, res.medoids);
+
+
 if __name__ == "__main__":
     unittest.main()
