### 2022-03-07 5d73397
Update documentation: Analog interface to sklearn.
