+++ README.md
@@ -61,6 +61,15 @@ c = kmedoids.fasterpam(distmatrix, 5)
 print("Loss is:", c.loss)
 ```
 
+###Using sklearn syntax
+
+```
+   import kmedoids
+   km = kmedoids.KMedoids(5, method='fasterpam')
+   c = km.fit(distmatrix)
+   print("Loss is:", c.inertia_)
+```
+
 ### MNIST (10k samples)
 
 ```
