### 2022-03-07 d6010a8
Enable syntax highlighting in code samples
