### 2022-03-08 2cca30f
Update sklearn-compatible API
