### 2022-03-08 f9f1a8c
Add fit_predict
