+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 kmedoids
+scikit-learn
