### 2022-03-27 baadf41
add metric_params, further improve documentation
