+++ kmedoids/__init__.py
@@ -15,8 +15,8 @@ References:
 | Erich Schubert, Peter J. Rousseeuw
 | Fast and Eager k-Medoids Clustering:
 | O(k) Runtime Improvement of the PAM, CLARA, and CLARANS Algorithms
-| Information Systems (101), 2021, 101804  
-| <https://doi.org/10.1016/j.is.2021.101804> (open access)
+| Information Systems (101), 2021, 101804
+| https://doi.org/10.1016/j.is.2021.101804 (open access)
 
 | Erich Schubert, Peter J. Rousseeuw:
 | Faster k-Medoids Clustering: Improving the PAM, CLARA, and CLARANS Algorithms
@@ -110,7 +110,7 @@ def fasterpam(diss, medoids, max_iter=100, init="random", random_state=None, n_c
 	| Fast and Eager k-Medoids Clustering:
 	| O(k) Runtime Improvement of the PAM, CLARA, and CLARANS Algorithms
 	| Information Systems (101), 2021, 101804
-	| <https://doi.org/10.1016/j.is.2021.101804> (open access)
+	| https://doi.org/10.1016/j.is.2021.101804 (open access)
 
 	| Erich Schubert, Peter J. Rousseeuw:
 	| Faster k-Medoids Clustering: Improving the PAM, CLARA, and CLARANS Algorithms
@@ -210,7 +210,7 @@ def fastpam1(diss, medoids, max_iter=100, init="random", random_state=None):
 	| Fast and Eager k-Medoids Clustering:
 	| O(k) Runtime Improvement of the PAM, CLARA, and CLARANS Algorithms
 	| Information Systems (101), 2021, 101804
-	| <https://doi.org/10.1016/j.is.2021.101804> (open access)
+	| https://doi.org/10.1016/j.is.2021.101804 (open access)
 
 	| Erich Schubert, Peter J. Rousseeuw:
 	| Faster k-Medoids Clustering: Improving the PAM, CLARA, and CLARANS Algorithms
@@ -447,10 +447,9 @@ def silhouette(diss, labels, samples=False, n_cpu=-1):
 	raise ValueError("Input data not supported. Use a numpy array of floats.")
 
 
+# This is a hack to make sklearn an optional dependency only:
 try:
 	from sklearn.base import BaseEstimator, ClusterMixin, TransformerMixin
-
-
 	class SKLearnClusterer(BaseEstimator, ClusterMixin, TransformerMixin):
 		pass
 except ImportError:
@@ -458,16 +457,41 @@ except ImportError:
 
 
 class KMedoids(SKLearnClusterer):
-	"""Provides sklearn-compatible API.
+	"""K-Medoids Clustering using PAM and FasterPAM (sklearn-compatible API).
+
+	References:
+
+	| Erich Schubert, Peter J. Rousseeuw
+	| Fast and Eager k-Medoids Clustering:
+	| O(k) Runtime Improvement of the PAM, CLARA, and CLARANS Algorithms
+	| Information Systems (101), 2021, 101804
+	| https://doi.org/10.1016/j.is.2021.101804 (open access)
+
+	| Erich Schubert, Peter J. Rousseeuw:
+	| Faster k-Medoids Clustering: Improving the PAM, CLARA, and CLARANS Algorithms
+	| In: 12th International Conference on Similarity Search and Applications (SISAP 2019), 171-187.
+	| https://doi.org/10.1007/978-3-030-32047-8_16
+	| Preprint: https://arxiv.org/abs/1810.05691
+
+	| Leonard Kaufman, Peter J. Rousseeuw:
+	| Clustering by means of medoids.
[…]
