+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
 name = "kmedoids"
-version = "0.2.2"
+version = "0.3.0"
 authors = ["Erich Schubert <erich.schubert@tu-dortmund.de>", "Lars Lenssen <lars.lenssen@tu-dortmund.de>"]
 description = "k-Medoids clustering with the FasterPAM algorithm"
 homepage = "https://github.com/kno10/python-kmedoids"
@@ -14,14 +14,17 @@ name = "kmedoids"
 crate-type = ["cdylib"]
 
 [dependencies]
-rustkmedoids = { version = "0.2.0", package = "kmedoids", git = "https://github.com/kno10/rust-kmedoids" }
-numpy = "0.15"
+rustkmedoids = { version = "0.3.0", package = "kmedoids", git = "https://github.com/kno10/rust-kmedoids" }
+numpy = "0.16"
 ndarray = "0.15"
 rand = "0.8"
 rayon = "1.5"
+# crossbeam is transitive via rayon, avoid a yanked version warning:
+# with rayon-core 1.9.2 this can likely be removed
+crossbeam-utils = {version = "0.8.8", optional = true}
 
 [dependencies.pyo3]
-version = "0.15.1"
+version = "0.16.2"
 features = ["extension-module"]
 
 [package.metadata.maturin]
