### 2022-04-05 e4f9154
Fix _check_medoids method
`_check_medoids` calls `warnings` but doesn't import the module. This
fix adds the import statement within the function scope.
