### 2022-04-05 33dd58c
Fix random init for duplicates
