### 2022-05-09 6dcbbd1
Amend installation instructions for non-Linux systems
Prebuilt packages are only for manylinux systems.
