### 2022-06-02 #4183 @timClicks
Re performance claims - the paper only claims to be faster than the upstream Java implementation. I don't think that the paper should be refused because it hasn't been benchmarked against all _k_-medoids/PAM implementations. Accordingly, I have marked the item as satisfied in my review.
