### 2022-06-02 kno10/python-kmedoids#10 @timClicks
Contribution Guidelines
Hello, I am one of the reviewers assigned to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4183). The journal's submission requirements require guidelines for for contributors.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This could take the shape of expanding the README or creating specific documentation. Here is some example text:

```markdown
## Contributing to `python-kmedoids`

Third-party contributions are welcome. Please use pull requests to submit patches.

## Reporting issues

Please report errors as an issue within the repository's issue tracker.

## Support requests

If you need help, please submit an issue within the repository's issue tracker.
```

Please note that you should create equivalent documentation within the Rust repository as well.
