### 2022-06-02 kno10/python-kmedoids#11 @timClicks
Statement of Need
Your paper requires a specific section header entitled "Statement of Need". Please see the reviewing guidelines:

>  A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
