### 2022-06-02 #4183 @timClicks
**Re "a statement of need"** The authors describe high performance + ease of use as their motivations. This satisfies the requirements within the Documentation section of the review, but unfortunately not the Software Paper section. It requires a paragraph entitled "Statement of Need". I have submitted an [issue in the Python repository](https://github.com/kno10/python-kmedoids/issues/11).

**Re Community Guidelines** I could not find contributor documentation when I looked. I have submitted an issue in the [Python repository](https://github.com/kno10/python-kmedoids/issues/10) along with some example wording.
