### 2022-06-02 #4183 @timClicks
> My major comment is about the package installation, it gives me a lot of errors.

Installation worked flawlessly for me today.
