### 2022-06-02 #4183 @TahiriNadia
> > My major comment is about the package installation, it gives me a lot of errors.
>
> Installation worked flawlessly for me today.

I'm trying it now and I confirm that it's OK for me too.
