### 2022-06-02 #4183 @TahiriNadia
> I put my comment on this document via PR. My major comment is about the package installation, it gives me a lot of errors. I asked the authors to correct this point by managing a requirements file. However, the idea of the paper is original and very relevant since it compares the best-known classification algorithms (i.e. kmedoids) in different programming languages (i.e. java, python, Rust, C++) and in classical python libraries (i.e. biopython, sklearn_extra, by clustering). This paper will be of great interest to the scientific community, as the task of classification is ubiquitous in various spheres of society (imaging, marketing, linguistics and others). Speed is often a critical element, especially when performing simulations of large-scale projects. Along with better documentation of how to install, a user guide, and finally a requirements file, I highly recommend this article. => FIX by @kno10
