### 2022-06-02 #4183 @mikldk
@TahiriNadia:  Can you confirm that you have finished the review and recommend that this paper is now published?

@timClicks: Thanks for [specifying the remaining items](https://github.com/openjournals/joss-reviews/issues/4183#issuecomment-1144396584) for @kno10
