### 2022-06-02 #4183 @lale1009
>

Hi @timClicks,
we benchmarked the performance against different implementations of kmedoids in different languages:
