### 2022-06-03 kno10/python-kmedoids#10 @lale1009
Done in 569e5855ef8390ed23f34ab059f3cf9f7be5a924
