### 2022-06-09 kno10/python-kmedoids#11 @lale1009
Section was added by @kno10.

4806c9e2a1c43e65d9620a34e369bc77ad005433
