### 2022-06-09 #4183 @lale1009
@mikldk and @timClicks: We have finished the [remaining items](https://github.com/openjournals/joss-reviews/issues/4183#issuecomment-1144396584)
