+++ paper.md
@@ -45,7 +45,7 @@ The implementation can be used with arbitrary dissimilarities and distances, as 
 To make the recent algorithmic improvements to k-medoids clustering available to a wider audience,
 we made an easy to use package available to the Rust and Python communities, to enable researchers
 to easily explore k-medoids clustering on their data sets, which so far is not available for example
-in the popular package scikit-learn (to which we include a compatible API).
+in the popular package scikit-learn (we include a compatible API).
 
 We chose Rust for the core functionality because of its high reliability, security, and performance,
 and a Python wrapper for ease of use. Both parts are documented following community best practice
