### 2022-06-15 1b24f4d
github action: build/upload as two steps
