### 2022-06-15 b1f077e
enable tests in the github actions build
