+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
 name = "kmedoids"
-version = "0.3.2"
+version = "0.3.3"
 authors = ["Erich Schubert <erich.schubert@tu-dortmund.de>", "Lars Lenssen <lars.lenssen@tu-dortmund.de>"]
 description = "k-Medoids clustering with the FasterPAM algorithm"
 homepage = "https://github.com/kno10/python-kmedoids"
@@ -14,7 +14,7 @@ name = "kmedoids"
 crate-type = ["cdylib"]
 
 [dependencies]
-rustkmedoids = { version = "0.3.2", package = "kmedoids", git = "https://github.com/kno10/rust-kmedoids" }
+rustkmedoids = { version = "0.3.3", package = "kmedoids", git = "https://github.com/kno10/rust-kmedoids" }
 numpy = "0.16"
 ndarray = "0.15"
 rand = "0.8"
