### 2022-07-06 #4183 @kno10
@mikldk: I have archived a snapshot of the Rust and Python parts version 0.3.3 at Zenodo as:
Schubert, Erich, & Lenssen, Lars. (2022). Fast k-medoids Clustering in Rust and Python (0.3.3). Zenodo. https://doi.org/10.5281/zenodo.6802320
The version number is 0.3.3 (some small bug fixes were integrated, and the package is now automatically built for more architectures to make it easier to install).
I do not have further changes to the paper itself.
