### 2022-07-07 #4183 @danielskatz
👋 Hi - I'm the AEiC working on finalizing the publication of this submission - I have one minor change in the paper in https://github.com/kno10/python-kmedoids/pull/15 that should be merged, then we can proceed.
